Creates external qr to human class
This class is used to hold a session with the app. An initial image is loaded and with each call, a volume level is returned. -1 is returned if the volume level shouldnt be changed. Also, this commit adds a flask post request handler to receive images so that it can be used from the ios platform without porting over any opencv code.